Create ParserScanner
This Java code accepts a input file which contains MiniLOL program. It first identifies all the keywords and identifiers used, that is Lexical analysis (Parser) and the next part of the code is to check of syntax which is known as Syntax analysis (Scanner). I have used Deterministic Finite Automata to implement this code.